+++ src/main/scala/io/viash/Main.scala
@@ -110,6 +110,14 @@ object Main {
           configs = configs,
           cli.namespace.list.format()
         )
+      case List(cli.namespace, cli.namespace.exec) =>
+        val configs = readConfigs(cli.namespace.exec, applyPlatform = false)
+        ViashNamespace.exec(
+          configs = configs,
+          command = cli.namespace.exec.cmd(),
+          dryrun = cli.namespace.exec.dryrun()
+        )
+
       case List(cli.config, cli.config.view) =>
         val config = Config.read(
           configPath = cli.config.view.config(),